requirements.txt
Flask==1.1.1
gunicorn==20.0.4
numpy==1.17.4
dash==1.13.4
dash_bootstrap_components==0.10.3
flask_caching==1.9.0
torch==1.8.1
torchvision==0.9.1
requests==2.24.0
matplotlib
Pillow==9.0.1
opencv-python==4.5.5.64
fast-dash==0.1.5